moving files, fixing REST structure to a proper one
