+++ settings.gradle
@@ -1,5 +1,6 @@
 plugins {
     id("com.gradle.enterprise") version("3.9")
+    id("org.gradle.toolchains.foojay-resolver-convention") version("0.4.0")
 }
 
 gradleEnterprise {
