+++ src/main/java/tech/ignitr/habitus/service/habit/HabitService.java
@@ -2,7 +2,7 @@ package tech.ignitr.habitus.service.habit;
 
 import org.springframework.http.ResponseEntity;
 import tech.ignitr.habitus.data.habit.Habit;
-import tech.ignitr.habitus.web.habit.HabitRequest;
+import tech.ignitr.habitus.web.habits.HabitRequest;
 
 import java.util.List;
 import java.util.UUID;
