+++ src/main/java/tech/ignitr/habitus/web/auth/AuthController.java
@@ -1,0 +1,30 @@
+package tech.ignitr.habitus.web.auth;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import tech.ignitr.habitus.data.user.User;
+import tech.ignitr.habitus.security.AuthService;
+import tech.ignitr.habitus.security.AuthenticationResponse;
+import tech.ignitr.habitus.service.user.UserService;
+import tech.ignitr.habitus.web.users.UserRequest;
+
+import java.util.UUID;
+
+@RestController
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final String baseURL = "/api";
+    private final AuthService authService;
+
+    @GetMapping(path=baseURL+"/auth/login")
+    public ResponseEntity<AuthenticationResponse> loginUser(@RequestBody LoginRequest request){
+        return authService.authenticateUser(request);
+    }
+
+    @PostMapping(path=baseURL+"/auth/register")
+    public ResponseEntity<AuthenticationResponse> putUser(@RequestBody RegisterRequest request){
+        return authService.registerUser(request);
+    }
+}
