+++ src/main/java/tech/ignitr/habitus/web/users/UserRequest.java
@@ -1,4 +1,4 @@
-package tech.ignitr.habitus.web.user;
+package tech.ignitr.habitus.web.users;
 
 import lombok.Getter;
